requirements.txt
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
fastapi>=0.85.0
slowapi
gradio
python-multipart
banana-dev

# We have to return images
Pillow